reorganise taint classes into a seperate file
